Add decorators for a2a and agent classes
