+++ package.json
@@ -33,7 +33,7 @@
     "example:streaming-example-workflow": "ts-node src/examples/streaming-example-workflow.ts",
     "example:streaming-example-team": "ts-node src/examples/streaming-example-team.ts",
     "example:web-page-content-example": "ts-node src/examples/web-page-content-example.ts",
-    "example:a2a-server": "ts-node src/examples/a2a-server-example.ts",
+    "example:a2a-server": "ts-node src/examples/a2a-example/a2a-server-example.ts",
     "example:a2a-team": "ts-node src/examples/a2a-team-example.ts"
   },
   "keywords": [
@@ -47,21 +47,20 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@microsoft/fetch-event-source": "^2.0.1",
-    "@modelcontextprotocol/sdk": "^1.12.0",
-    "@types/node": "^22.15.21",
-    "axios": "^1.9.0",
-    "cheerio": "^1.0.0",
-    "dotenv": "^16.5.0",
-    "event-source-polyfill": "^1.0.31",
-    "eventsource": "^4.0.0",
-    "token.js": "^0.7.1",
-    "typescript": "^5.8.3",
-    "yaml": "^2.8.0",
-    "zod": "^3.25.23"
+    "@microsoft/fetch-event-source": "2.0.1",
+    "@modelcontextprotocol/sdk": "1.12.0",
+    "axios": "1.9.0",
+    "cheerio": "1.0.0",
+    "dotenv": "16.5.0",
+    "event-source-polyfill": "1.0.31",
+    "eventsource": "4.0.0",
+    "token.js": "0.7.1",
+    "typescript": "5.8.3",
+    "yaml": "2.8.0",
+    "zod": "3.25.23"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.9.4",
+    "@biomejs/biome": "^1.9.4",
     "@bobvanderlinden/semantic-release-pull-request-analyzer": "^1.0.4",
     "@docusaurus/core": "^3.7.0",
     "@docusaurus/preset-classic": "^3.7.0",
@@ -69,6 +68,7 @@
     "@semantic-release/git": "^10.0.1",
     "@types/eventsource": "^3.0.0",
     "@types/jest": "^29.5.14",
+    "@types/node": "22.15.21",
     "@types/react": "^19.1.5",
     "@types/react-dom": "^19.1.5",
     "jest": "^29.7.0",
