+++ src/a2a/decorators.ts
@@ -3,6 +3,7 @@ import { RemoteA2AAgent } from "./client/RemoteA2AAgent";
 import type { A2AClientOptions } from "./client/types";
 import { A2AServer } from "./server/A2AServer";
 import type { A2AServerOptions } from "./server/types";
+// import { agent } from "../core/decorators"; // If you want to re-export
 
 /**
  * Class decorator to make an Agent class use a remote A2A agent via the A2A protocol.
